Animation 'Y' axis revised
The Player was dropping about a half meter when the game started. Solution, the animations 'Y' access was a half meter too low throughout the game and so was starting where I told it to. Adjusting all animations corrected this problem. Movement through the level is correct now with not dropping of the player.